+++ lib/basketball/scheduling.rb
@@ -1,12 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'scheduling/calendar_serializer'
-require_relative 'scheduling/conference'
-require_relative 'scheduling/coordinator'
-require_relative 'scheduling/division'
-require_relative 'scheduling/league'
-require_relative 'scheduling/league_serializer'
-require_relative 'scheduling/team'
+require_relative 'scheduling/cli'
 
 module Basketball
   module Scheduling