+++ lib/basketball/scheduling/preseason_game.rb
@@ -3,6 +3,9 @@
 module Basketball
   module Scheduling
     class PreseasonGame < Game
+      def to_s
+        "#{super} (preseason)"
+      end
     end
   end
 end